frontend/requirements.txt
streamlit>=1.22.0
requests>=2.28.2
python-dotenv>=1.0.0
pandas>=1.5.3
plotly>=5.13.1
flask>=2.2.3
flask-cors>=3.0.10